requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==2.0.5
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==2.0.0.dev1
django-formtools==2.1
django-import-export==1.0.1
django-reversion==2.0.13
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
idna==2.6
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.6
olefile==0.44
openpyxl==2.5.3
Pillow==5.1.0
PyJWT==1.6.3
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
XlsxWriter==1.0.5
xlwt==1.3.0
